Breaking change on how variables are defined
Move the automatic syncing definitions to the actual array definition
so two separate lists don't need to be maintained.

Move the responsibility of naming and creating the primary content to
the ansible scripts.

Move some email variables to the top level so they don't need to be
repeated for every array.